chat/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7